crawling/requirements.txt
pandas == 1.5.0
datetime == 4.9
boto3 == 1.26.38
python-dotenv
requests
bs4